Added an Not Connected overlay
Not connected overlay is an overlay which covers tabs
when app isn't connected to controller.
Unfortunately it needs a little refactoring.
An additional onTabSelected and onTabUnselected methods
has been written to separate starting thread from Bluetooth
Controller and Tab listener. But i don't like this. Probably code
needs to be inspected and rewritten if needed.